+++ scripts/git.cmake
@@ -1,4 +1,0 @@
-find_package(Git)
-if(Git_FOUND)
-    message("Git found: ${GIT_EXECUTABLE}")
-endif()