+++ src/common.hpp
@@ -1,9 +1,0 @@
-#ifndef COMMON_H
-#define COMMON_H
-
-namespace common
-{
-	inline std::string_view version = "1.4.0";
-}
-
-#endif //COMMON_H