+++ src/custom_layers/ConfirmationLayer.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include "../pch.hpp"
 #include "base/BaseCustomLayer.hpp"
 
 class ConfirmationAlertLayer;